+++ src/flask_session/__init__.py
@@ -9,7 +9,7 @@ from .sessions import (
     SqlAlchemySessionInterface,
 )
 
-__version__ = "0.6.0rc1"
+__version__ = "0.6.1"
 
 
 class Session: