+++ src/flask_session/sessions.py
@@ -137,9 +137,10 @@ class ServerSideSessionInterface(SessionInterface, ABC):
     def fetch_session(self, sid):
         raise NotImplementedError()
 
-    def regenerate_sid(self):
+    def regenerate_sid(self, session):
         sid = self._generate_sid(self.sid_length)
-        return self.session_class(sid=sid, permanent=self.permanent)
+        session.sid = sid
+        session.modified = True
 
 
 class RedisSessionInterface(ServerSideSessionInterface):
@@ -181,6 +182,9 @@ class RedisSessionInterface(ServerSideSessionInterface):
                 return self.session_class(session_data, sid=sid)
             except pickle.UnpicklingError:
                 return self.session_class(sid=sid, permanent=self.permanent)
+        else:
+            # If there is no saved session on the server, generate a new session id to mitigate session fixation vulnerability
+            sid = self._generate_sid(self.sid_length)
 
         # If the saved session  does not exist, create a new session
         return self.session_class(sid=sid, permanent=self.permanent)
@@ -285,6 +289,9 @@ class MemcachedSessionInterface(ServerSideSessionInterface):
                 return self.session_class(session_data, sid=sid)
             except pickle.UnpicklingError:
                 return self.session_class(sid=sid, permanent=self.permanent)
+        else:
+            # If there is no saved session on the server, generate a new session id to mitigate session fixation vulnerability
+            sid = self._generate_sid(self.sid_length)
 
         # If the saved session  does not exist, create a new session
         return self.session_class(sid=sid, permanent=self.permanent)
@@ -371,6 +378,9 @@ class FileSystemSessionInterface(ServerSideSessionInterface):
         # If the saved session exists and has not auto-expired, load the session data from the item
         if item is not None:
             return self.session_class(item, sid=sid)
+        else:
+            # If there is no saved session on the server, generate a new session id to mitigate session fixation vulnerability
+            sid = self._generate_sid(self.sid_length)
 
         # If the saved session  does not exist, create a new session
         return self.session_class(sid=sid, permanent=self.permanent)
@@ -483,6 +493,9 @@ class MongoDBSessionInterface(ServerSideSessionInterface):
                 return self.session_class(session_data, sid=sid)
             except pickle.UnpicklingError:
                 return self.session_class(sid=sid, permanent=self.permanent)
+        else:
+            # If there is no saved session on the server, generate a new sid to mitigate session fixation vulnerability
+            sid = self._generate_sid(self.sid_length)
 
         # If the saved session does not exist, create a new session
         return self.session_class(sid=sid, permanent=self.permanent)
@@ -649,6 +662,10 @@ class SqlAlchemySessionInterface(ServerSideSessionInterface):
                 return self.session_class(session_data, sid=sid)
             except pickle.UnpicklingError:
                 return self.session_class(sid=sid, permanent=self.permanent)
+        else:
+            # If there is no saved session on the server, generate a new session id to mitigate session fixation vulnerability
+            sid = self._generate_sid(self.sid_length)
+
         return self.session_class(sid=sid, permanent=self.permanent)
 
     def save_session(self, app, session, response):